requirements.txt
openai
python-dotenv
sentence-transformers
sounddevice 
numpy
faster-whisper 
torch
pyttsx3
